Do not include Movement in Booking output

+++ BasicAccounting/src/test/java/be/dafke/BasicAccounting/Objects/MovementTest.java
@@ -25,7 +25,7 @@ public class MovementTest {
         int id = movement.getId();
 
         movement = new Movement(AMOUNT, false);
-        assertEquals(id+1,movement.getId());
+        assertEquals(id+1,movement.getId().intValue());
     }
 
     @Test
